fleetManagement/requirements.txt
asgiref==3.7.2
astroid==3.0.2
dill==0.3.7
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==7.4.4
pytest-django==4.7.0
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.3
psycopg2==2.9.3
